+++ cmd/korrel8r/web_test.go
@@ -88,8 +88,6 @@ func TestMain_server_insecure(t *testing.T) {
 {"name":"metric"},
 {"name":"mock","stores":[{"domain":"mock", "mockData":"testdata/mock_store.yaml"}]},
 {"name":"netflow"},
-{"name":"otellog"},
-{"name":"podlog"},
 {"name":"trace"}
 ]`, "GET", u, "")
 }
@@ -106,8 +104,6 @@ func TestMain_server_secure(t *testing.T) {
 {"name":"metric"},
 {"name":"mock","stores":[{"domain":"mock", "mockData":"testdata/mock_store.yaml"}]},
 {"name":"netflow"},
-{"name":"otellog"},
-{"name":"podlog"},
 {"name":"trace"}
 ]`,
 		"GET", u, "")