+++ internal/pkg/test/domain/fixture.go
@@ -50,7 +50,7 @@ func (f *Fixture) ClusterEngine(t testing.TB) *engine.Engine {
 	// TODO review test logic for cluster vs. no-cluster tests.
 	t.Helper()
 	if f.SkipCluster {
-		t.Skipf("Skip: domain %v skipping openshift tests", f.Query.Class().Domain())
+		t.Skipf("Skip: domain %v skipping cluster tests", f.Query.Class().Domain())
 	}
 	test.SkipIfNoCluster(t)