+++ pkg/domains/k8s/k8s.go
@@ -2,6 +2,27 @@
 
 // Package k8s implements [Kubernetes] resources stored in a Kube API server.
 //
+// # Class
+//
+// Class represents a kind of kubernetes resource.
+// The format of a class name is: "k8s:KIND[.VERSION][.GROUP]".
+//
+// Missing VERSION implies "v1", if present VERSION must follow the [Kubernetes version patterns].
+// Missing GROUP implies the core group.
+//
+// Examples: `k8s:Pod`, `ks8:Pod/v1`, `k8s:Deployment.apps`, `k8s:Deployment.apps/v1`, `k8s:Route.route.openshift.io/v1`
+//
+// # Object
+//
+// Object represents a kubernetes resource as a map, map keys are serialized field names.
+// Rule templates should use the JSON (lowerCase) field names, NOT the UpperCase Go struct field names.
+//
+// # Query
+//
+// Example:
+//
+//	k8s:Pod.v1:{"namespace":"openshift-cluster-version","name":"cluster-version-operator-8d86bcb65-btlgn"}
+//
 // # Store
 //
 // The k8s domain automatically connects to the current cluster (as determined by kubectl),
@@ -11,6 +32,7 @@
 //		  domain: k8s
 //
 // [Kubernetes]: https://kubernetes.io/docs/concepts/overview/
+// [Kubernetes version patterns]: https://kubernetes.io/docs/tasks/extend-kubernetes/custom-resources/custom-resource-definition-versioning/#version-priority
 package k8s
 
 import (
@@ -39,26 +61,10 @@ import (
 // Domain for Kubernetes resources stored in a Kube API server.
 var Domain = domain{}
 
-// Class represents a kind of kubernetes resource.
-// The format of a class name is: "k8s:KIND[.VERSION][.GROUP]".
-//
-// Missing VERSION implies "v1", if present VERSION must follow the [Kubernetes version patterns].
-// Missing GROUP implies the core group.
-//
-// Examples: `k8s:Pod`, `ks8:Pod/v1`, `k8s:Deployment.apps`, `k8s:Deployment.apps/v1`, `k8s:Route.route.openshift.io/v1`
-//
-// [Kubernetes version patterns]: https://kubernetes.io/docs/tasks/extend-kubernetes/custom-resources/custom-resource-definition-versioning/#version-priority
 type Class schema.GroupVersionKind
 
-// Object represents a kubernetes resource as a map, map keys are serialized field names.
-// Rule templates should use the JSON (lowerCase) field names, NOT the UpperCase Go struct field names.
 type Object = map[string]any
 
-// Query struct for a Kubernetes query.
-//
-// Example:
-//
-//	k8s:Pod.v1:{"namespace":"openshift-cluster-version","name":"cluster-version-operator-8d86bcb65-btlgn"}
 type Query struct {
 	Selector
 	class Class // class is the underlying k8s.Class object. Implied by query name prefix.