+++ pkg/korrel8r/impl/query.go
@@ -6,7 +6,6 @@ import (
 	"fmt"
 
 	"github.com/korrel8r/korrel8r/pkg/korrel8r"
-	"sigs.k8s.io/yaml"
 )
 
 // ParseQuery parses a query string into class and data parts.
@@ -32,7 +31,7 @@ func UnmarshalQueryString[T any](domain korrel8r.Domain, query string) (c korrel
 	if err != nil {
 		return nil, data, err
 	}
-	err = yaml.UnmarshalStrict([]byte(qs), &data)
+	err = Unmarshal([]byte(qs), &data)
 	if err != nil {
 		return nil, data, fmt.Errorf("invalid query: %w: %v", err, qs)
 	}